+++ ref/hash.h
@@ -2,6 +2,7 @@
 #define SPX_HASH_H
 
 #include <stdint.h>
+#include <stddef.h>
 
 #include "context.h"
 #include "params.h"
@@ -13,13 +14,13 @@ void prf_addr(unsigned char *out, const spx_ctx *ctx,
 #define gen_message_random SPX_NAMESPACE(gen_message_random)
 void gen_message_random(unsigned char *R, const unsigned char *sk_prf,
                         const unsigned char *optrand,
-                        const unsigned char *m, unsigned long long mlen,
+                        const unsigned char *m, size_t mlen,
                         const spx_ctx *ctx);
 
 #define hash_message SPX_NAMESPACE(hash_message)
 void hash_message(unsigned char *digest, uint64_t *tree, uint32_t *leaf_idx,
                   const unsigned char *R, const unsigned char *pk,
-                  const unsigned char *m, unsigned long long mlen,
+                  const unsigned char *m, size_t mlen,
                   const spx_ctx *ctx);
